+++ test/dataframe.jl
@@ -613,6 +613,17 @@ end
         @test delete!(df, v) == DataFrame(x=[2, 3])
         @test x == [2, 3]
     end
+
+    for inds in (1, [1], [true, false])
+        df = DataFrame(x1=[1, 2])
+        df.x2 = df.x1
+        @test delete!(df, inds) === df
+        @test df == DataFrame(x1=[2], x2=[2])
+    end
+
+    df = DataFrame(a=1, b=2)
+    push!(df.b, 3)
+    @test_throws AssertionError delete!(df, 1)
 end
 
 @testset "describe" begin